requirements.txt
brainglobe-atlasapi == 2.0.3
torch == 2.2.0
torchvision == 0.17.0
scikit-learn == 1.4.0
opencv-python == 4.9.0.80
numpy == 1.26.4
matplotlib == 3.8.3
tifffile == 2024.2.12
Pillow == 10.2.0
umap-learn == 0.5.5
tqdm == 4.66.2
fastapi == 0.109.2
uvicorn == 0.27.1
python-multipart == 0.0.9
aiofiles == 23.2.1
pytest == 8.0.2
httpx == 0.27.0